feat: Implement reusable Glass UI framework and integrate image cropper
- Resolved a startup crash by mapping `Palette` ARGB variables as `Int` instead of `Long` to avoid Compose color-space out-of-bounds exceptions.
- Prevented potential OOM exceptions in `WallpaperRepository` by safely downsampling the user's gallery `InputStream` via `inJustDecodeBounds` size calculations before saving it.
- Integrated `Android-Image-Cropper` into the `WallpaperScreen` to allow exact dimension wallpaper framing.
- Built a robust, customizable `GlassUI.kt` module exposing `GlassContainer`, `GlassCard`, and `GlassBackground` composables.
- Integrated the frosted glass effect deeply into the `AppDrawer` and `FolderItem` views to match the required OS aesthetic.

+++ app/src/main/java/com/eclipse/launcher/data/repository/DynamicThemeRepositoryImpl.kt
@@ -45,9 +45,9 @@ class DynamicThemeRepositoryImpl @Inject constructor(
 
         val palette = Palette.from(bitmap).generate()
 
-        val dominantColor = palette.getDominantColor(Color.BLACK).toLong()
-        val secondaryColor = palette.getMutedColor(dominantColor.toInt()).toLong()
-        val accentColor = palette.getVibrantColor(dominantColor.toInt()).toLong()
+        val dominantColor = palette.getDominantColor(Color.BLACK)
+        val secondaryColor = palette.getMutedColor(dominantColor)
+        val accentColor = palette.getVibrantColor(dominantColor)
 
         // Calculate derivatives
         val backgroundColor = adjustAlpha(dominantColor, 0.9f) // Slight transparency for the wallpaper
@@ -55,11 +55,11 @@ class DynamicThemeRepositoryImpl @Inject constructor(
         val dockColor = adjustAlpha(dominantColor, 0.7f) // Slightly darker than background
 
         // Ensure contrast for icons
-        val isDarkBackground = ColorUtils.calculateLuminance(dominantColor.toInt()) < 0.5
+        val isDarkBackground = ColorUtils.calculateLuminance(dominantColor) < 0.5
         val iconTintColor = if (isDarkBackground) {
-            Color.WHITE.toLong()
+            Color.WHITE
         } else {
-            Color.BLACK.toLong()
+            Color.BLACK
         }
 
         DynamicThemeColors(
@@ -73,11 +73,11 @@ class DynamicThemeRepositoryImpl @Inject constructor(
         )
     }
 
-    private fun adjustAlpha(color: Long, factor: Float): Long {
-        val alpha = Math.round(Color.alpha(color.toInt()) * factor)
-        val r = Color.red(color.toInt())
-        val g = Color.green(color.toInt())
-        val b = Color.blue(color.toInt())
-        return Color.argb(alpha, r, g, b).toLong()
+    private fun adjustAlpha(color: Int, factor: Float): Int {
+        val alpha = Math.round(Color.alpha(color) * factor)
+        val r = Color.red(color)
+        val g = Color.green(color)
+        val b = Color.blue(color)
+        return Color.argb(alpha, r, g, b)
     }
 }
