+++ app/src/main/java/com/eclipse/launcher/domain/model/DynamicThemeColors.kt
@@ -1,11 +1,11 @@
 package com.eclipse.launcher.domain.model
 
 data class DynamicThemeColors(
-    val dominantColor: Long? = null,
-    val secondaryColor: Long? = null,
-    val accentColor: Long? = null,
-    val backgroundColor: Long? = null,
-    val widgetColor: Long? = null,
-    val dockColor: Long? = null,
-    val iconTintColor: Long? = null
+    val dominantColor: Int? = null,
+    val secondaryColor: Int? = null,
+    val accentColor: Int? = null,
+    val backgroundColor: Int? = null,
+    val widgetColor: Int? = null,
+    val dockColor: Int? = null,
+    val iconTintColor: Int? = null
 )
